bug fixes, functionality and more

+++ include/MANEUVER.h
@@ -3,9 +3,9 @@
 #define steeringPin 25     // Pin for steering servo
 #define motorPin    26     // Pin for motor servo
 
-const float steeringOffset = 30; // Steering offset to prevent servo damage
-const float centerSteeringAngle = 90; // Center angle for steering (to account for joystick drift)
-const float centerSteeringTolerance = 3; // Tolerance for centering the steering (to account for joystick drift)
+const uint8_t steeringOffset = 30; // Steering offset to prevent servo damage
+const uint8_t centerSteeringAngle = 90; // Center angle for steering (to account for joystick drift)
+const uint8_t centerSteeringTolerance = 3; // Tolerance for centering the steering (to account for joystick drift)
 
 Servo absimaServo; // Servo object for steering
 Servo absimaMotor; // Servo object for motor control
@@ -27,7 +27,7 @@ void setupMANEUVER () {
 
 //==================================================================================/
 
-void maneuver(int16_t throttle, float steeringAngle){
+void maneuver(int16_t throttle, uint8_t steeringAngle){
     // Center steering angle if within tolerance
 
     if (abs(steeringAngle - centerSteeringAngle) <= centerSteeringTolerance) {
